+++ src/utils/index.js
@@ -3,3 +3,4 @@ export { default as unusedCityCards } from './unusedCityCards'
 export { default as unusedInfectionCards } from './unusedInfectionCards'
 export { default as collections } from './collections'
 export { default as init } from './initialGameState'
+export { default as setupLogic } from './setupLogic';
